fix: 11 bug revisi klien — DANA QR, estimasi hari, eye password, chat online/offline, hapus pesan, filter sidebar
- Payment: tambah QR barcode DANA + auto-generate via qrserver.com, warna area upload bukti
- Estimasi hari: cast (int) pada diffInDays() agar tidak tampil 13.999999
- Login & Register: tombol eye toggle show/hide password (Alpine x-data per field)
- Admin shipment form: styling warna orange gradient agar terlihat jelas
- Portfolio & Catalog CRUD: fix Storage::disk('public'), tambah mimes webp, perbaiki update() is_active
- Order status: permak → 'Selesai Dijahit', custom → 'Selesai Dibuat' via getDoneLabel()
- Chat: formatted_time dari server (fix timezone JS), adminOnline/userOnline seed dari PHP (fix flicker Offline)
- Chat: tombol hapus semua pesan (user: pesan sendiri, admin: seluruh percakapan)
- Admin chat sidebar: filter/cari nama pelanggan real-time
- Sidebar online dot update via updateSidebarUnread() polling

+++ resources/views/user/orders/show.blade.php
@@ -86,7 +86,7 @@
                 <p class="text-xs mt-1.5 font-semibold {{ $bannerConfig['title_color'] }}">
                     Estimasi selesai: {{ $estimatedDone->format('d F Y') }}
                     <span class="font-normal {{ $bannerConfig['text_color'] }}">
-                        ({{ $estimatedDone->isFuture() ? 'sekitar ' . now()->diffInDays($estimatedDone) . ' hari lagi' : 'sudah melewati estimasi' }})
+                        ({{ $estimatedDone->isFuture() ? 'sekitar ' . (int) now()->diffInDays($estimatedDone) . ' hari lagi' : 'sudah melewati estimasi' }})
                     </span>
                 </p>
             @endif
